Error when try to export excel with total

Related Work Items: #71470

+++ Logitude/JustWebFreight/WebFreight.Web/Helpers/ExportToExcelHelper.cs
@@ -726,7 +726,7 @@ namespace WebFreight.Web.Helpers
                             string value = row[agColumn.Name].ToString();
                             if (!string.IsNullOrEmpty(value))
                             {
-                                excelTotals[excelTotalCount].Total += Int32.Parse(value);
+                                excelTotals[excelTotalCount].Total += Double.Parse(value);
                             }
                             excelTotalCount += 1;
                         }
